+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Compatibility:
 * Support the `resolv_timeout` and `fast_fallback` keyword arguments for `TCPSocket.new` (#3883, @eregon).
 * `require 'readline'` now uses either `Reline` or the `readline-ext` gem, same as CRuby (#3018, @eregon).
 * Update `Kernel#singleton_method` to implement Ruby 3.4 singleton ancestor lookup semantics (#3883, @nirvdrum).
+* `Process.spawn` now raises `Errno::EACCES` instead of `Errno::ENOENT` when the file exists but is not executable (#4176, @alessandro54).
 
 Performance: